Remove a bunch of unused junk
Simplify the ravings. This also removes the absurd abundance
of great items the Dreamsnatcher used to spawn in service of
The Raven performance. Maybe someday the raven himself will
reappear to strut about upon the stage and mock the player's
hopes and regrets. Until then he lives in a git tree...

Signed-off-by: Matt Helsley <matt.helsley@gmail.com>