Add crafting icon and placer prefab
We provide some crude wrappers to the Shipwrecked and Don't Starve
Together Recipe constructor because they changed the parameter list
from the plain/Reign of Giants versions. This in turn requires
checking for SW/RoG DLC and whether this is DST. We do that as
a sequence of wrappers rather than a combinatorial number of
if/then/else because perhaps someday all of these will be possible
to enable in the same game.

Signed-off-by: Matt Helsley <matt.helsley@gmail.com>