Tweak sanity accumulation in dreamsnatcher
Sanity accumulation was too low. Refactor it to make the code
easier to understand. Push events for accumulation and
loot dropping. Adjust the accumulation formula for NPCs so
we can produce nightmare fuel in reasonable amounts of time with
reasonable effort.

Also, escape a bunch of the ravings in the hopes it gets rid
of some of the misprinted lines.

Signed-off-by: Matt Helsley <matt.helsley@gmail.com>